Move some tests to the "abstract" module
