+++ tests/matchers/test_abstract.py
@@ -1,7 +1,7 @@
 import pytest
 from full_match import match
 
-from sigmatch import FunctionSignatureMatcher, PossibleCallMatcher, IncorrectArgumentsOrderError
+from sigmatch import FunctionSignatureMatcher, PossibleCallMatcher, IncorrectArgumentsOrderError, SignatureMismatchError
 
 
 def test_match_not_callable(matcher_class):
@@ -244,3 +244,63 @@ def test_strict_match_random_generator_functions(matcher_class):
     assert matcher_class('.', '.', 'c', 'c2', '*', '**').match(function_10)
     assert matcher_class('.', '.', '.', 'c', 'c2', '*', '**').match(function_11)
     assert matcher_class('c', 'c2').match(function_12)
+
+
+def test_raise_exception_if_dismatch(matcher_class):
+    with pytest.raises(SignatureMismatchError):
+        matcher_class().match(lambda x: None, raise_exception=True)
+
+
+@pytest.mark.parametrize(
+    'options',
+    [
+        {},
+        {'raise_exception': False},
+    ],
+)
+def test_not_raise_exception_if_dismatch_and_flag_is_false(options, matcher_class):
+    assert not matcher_class().match(lambda x: None, **options)
+
+
+def test_it_works_with_class_based_callables(matcher_class):
+    class LocalCallable:
+        def __call__(self):
+            pass
+
+    assert matcher_class().match(LocalCallable)
+    assert not matcher_class('.').match(LocalCallable)
+
+
+def test_empty_class_as_callable(matcher_class):
+    class Kek:
+        pass
+
+    assert matcher_class().match(Kek)
+    assert not matcher_class('.').match(Kek)
+
+
+def test_class_with_init_as_callable(matcher_class):
+    class Kek:
+        def __init__(self, a, b, c):
+            pass
+
+    assert matcher_class('.', '.', '.').match(Kek)
+    assert not matcher_class().match(Kek)
+
+
+def test_class_with_call_dunder_object_is_callable(matcher_class):
+    class Kek:
+        def __call__(self, a, b, c):
+            pass
+
+    assert matcher_class('.', '.', '.').match(Kek())
+    assert not matcher_class().match(Kek())
+
+
+def test_check_method(matcher_class):
+    class Kek:
+        def kek(self, a, b, c):
+            pass
+
+    assert matcher_class('.', '.', '.').match(Kek().kek)
+    assert not matcher_class().match(Kek().kek)
