+++ tests/matchers/test_function_matcher.py
@@ -132,66 +132,3 @@ def test_random_wrong_generator_functions():
     assert FunctionSignatureMatcher('.', '.', 'c2', '*', '**').match(function_10) == False
     assert FunctionSignatureMatcher('.', '.', 'c2', '*', '**').match(function_11) == False
     assert FunctionSignatureMatcher('c').match(function_12) == False
-
-
-
-
-
-def test_raise_exception_if_dismatch():
-    with pytest.raises(SignatureMismatchError):
-        FunctionSignatureMatcher().match(lambda x: None, raise_exception=True)
-
-
-@pytest.mark.parametrize(
-    'options',
-    [
-        {},
-        {'raise_exception': False},
-    ],
-)
-def test_not_raise_exception_if_dismatch_and_flag_is_false(options):
-    assert FunctionSignatureMatcher().match(lambda x: None, **options) == False
-
-
-def test_it_works_with_class_based_callables():
-    class LocalCallable:
-        def __call__(self):
-            pass
-
-    assert FunctionSignatureMatcher().match(LocalCallable)
-    assert not FunctionSignatureMatcher('.').match(LocalCallable)
-
-
-def test_empty_class_as_callable():
-    class Kek:
-        pass
-
-    assert FunctionSignatureMatcher().match(Kek)
-    assert not FunctionSignatureMatcher('.').match(Kek)
-
-
-def test_class_with_init_as_callable():
-    class Kek:
-        def __init__(self, a, b, c):
-            pass
-
-    assert FunctionSignatureMatcher('.', '.', '.').match(Kek)
-    assert not FunctionSignatureMatcher().match(Kek)
-
-
-def test_class_with_call_dunder_object_is_callable():
-    class Kek:
-        def __call__(self, a, b, c):
-            pass
-
-    assert FunctionSignatureMatcher('.', '.', '.').match(Kek())
-    assert not FunctionSignatureMatcher().match(Kek())
-
-
-def test_check_method():
-    class Kek:
-        def kek(self, a, b, c):
-            pass
-
-    assert FunctionSignatureMatcher('.', '.', '.').match(Kek().kek)
-    assert not FunctionSignatureMatcher().match(Kek().kek)
